Address PR comments

+++ dnn-providers/integration-tests/tests/LayernormBwdGraphTestUtils.hpp
@@ -42,7 +42,7 @@ inline flatbuffers::FlatBufferBuilder
                             const std::vector<int64_t>& dbiasStrides,
                             const std::optional<std::vector<int64_t>>& meanStrides,
                             const std::optional<std::vector<int64_t>>& invVarianceStrides,
-                            const std::optional<double> epsilon,
+                            const std::optional<float> epsilon,
                             const int64_t normalizedDimCount,
                             const DataType dyDataType,
                             const DataType dxDataType,
@@ -72,12 +72,12 @@ inline flatbuffers::FlatBufferBuilder
             builder,
             epsilonUid.value(),
             "epsilon",
-            DataType::DOUBLE,
+            DataType::FLOAT,
             &epsilonDimsStrides,
             &epsilonDimsStrides,
             false,
-            TensorValue::Float64Value,
-            builder.CreateStruct(Float64Value(epsilon.value())).Union()));
+            TensorValue::Float32Value,
+            builder.CreateStruct(Float32Value(epsilon.value())).Union()));
     }
     if(meanUid.has_value() && meanDims.has_value() && meanStrides.has_value()
        && meanInvVarianceDataType.has_value())
@@ -143,7 +143,7 @@ inline flatbuffers::FlatBufferBuilder
                             const std::optional<int64_t> invVarianceUid,
                             const std::vector<int64_t>& ioDims,
                             const TensorLayout& layout,
-                            const double epsilon,
+                            const float epsilon,
                             const int64_t normalizedDimCount,
                             const DataType dyDataType,
                             const DataType dxDataType,
